\name{mice.A}
\alias{mice.A}
\title{Pedigree info for the mice dataset}
\description{

Is a numerator relationship matrix (1814 x 1814) computed
from a pedigree that traced back many generations.

}

\source{
  \url{http://gscan.well.ox.ac.uk} 
}

\references{
de los Campos G., H. Naya, D. Gianola, J. Crossa, A. Legarra, E. Manfredi, K. Weigel and J. Cotes. 2009.
Predicting Quantitative Traits with Regression Models for Dense Molecular Markers and Pedigree. \emph{Genetics} \bold{182}: 375-385.
}
\keyword{datasets}
