\name{mice.pheno}
\alias{mice.pheno}
\title{Phenotypical data for the mice dataset}
\description{

A data frame with pheotypical information related to obesity and 
biochemistry. 
The data frame has several columns: SUBJECT.NAME, Obesity.BMI, Obesity.BodyLength, 
Obesity.Date.Month, Obesity.Date.Year, Obesity.Date.Season,
Obesity.Date.StudyStartSeconds, Obesity.Date.Hour, Obesity.Date.StudyDay,          
GENDER, Obesity.EndNormalBW, CoatColour, CageDensity, Litter,                        
cage, Biochem.Albumin, Biochem.ALP, Biochem.ALT,                   
Biochem.AST, Biochem.Calcium, Biochem.Chloride, Biochem.Creatinine,
Biochem.Glucose, Biochem.HDL, Biochem.LDL, Biochem.Phosphorous,           
Biochem.Potassium, Biochem.Sodium, Biochem.Tot.Cholesterol,
Biochem.Tot.Protein, Biochem.Triglycerides, Biochem.Urea,                  
Biochem.EndNormalBW, Biochem.Date.StudyDay, Biochem.Date.Season,
Biochem.Date.Month, Biochem.Date.Year and Biochem.Age.  

The data can be downloaded from http://mtweb.cs.ucl.ac.uk/mus/www/GSCAN/HS_PHENOTYPES/.

}

\keyword{datasets}
