\name{mice}
\alias{mice}
\title{mice dataset}
\description{

The mice data comes from an experiment carried out to detect and locate QTLs for complex traits in a 
mice population (Valdar et al. 2006a; 2006b). This data has already been analyzed for comparing genome-assisted 
genetic evaluation methods (Legarra et al. 2008). The data file consists of 1814 individuals,  
each genotyped for 10,346 polymorphic markers. The information is related to 
obesity and biochemistry and other covariates related to the traits.

}

\usage{
  data(mice)
}

\format{

 Matrix mice.A contains the pedigree.
 The matrix mice.X contains the markes information and the map data.frame
 contains information for the genetic map.
 The data.frame mice.pheno contains phenotypical information.

}

\references{

Legarra A., Robert-Granie, E. Manfredi, and J. M. Elsen, 2008 Performance of genomic selection in mice. Genetics 180:611-618.

Valdar, W., L. C. Solberg, D. Gauguier, S. Burnett, P. Klenerman et al., 2006a Genome-wide genetic association of complex traits in heterogeneous stock mice. Nat. Genet. 38:879-887.

Valdar, W., L. C. Solberg, D. Gauguier, W. O. Cookson, J. N. P. Rawlis et al., 2006b Genetic and environmental effects on complex traits in mice. Genetics, 174:959-984.

}


\keyword{datasets}
