\name{simulated3t.pheno}
\alias{simulated3t.pheno}
\title{Phenotypical data for simulated dataset with 3 traits}
\description{

A matrix with 8000 rows and 6 columns.
Columns 1 to 3 corresponds to simulated phenotypes for 3 traits,
columns 4-6 corresponds to true simulated genetic values for 3 traits.

}

\keyword{datasets}
