\name{wheat.A}
\alias{wheat.A}
\title{Pedigree info for the wheat dataset}
\description{

Is a numerator relationship matrix (599 x 599) computed
from a pedigree that traced back many generations.
This relationship matrix was derived using the
Browse application of the International Crop Information
System (ICIS), as described in  
\url{http://repository.cimmyt.org/xmlui/bitstream/handle/10883/3488/72673.pdf} (McLaren \emph{et al.} 2000, 2005).

}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\references{

McLaren, C. G., L. Ramos, C. Lopez, and W. Eusebio. 2000. ``Applications of the geneaology manegment system.''
In \emph{International Crop Information System. Technical  Development Manual, version VI}, edited by McLaren, C. G., J.W. White
and P.N. Fox. pp. 5.8-5.13. CIMMyT, Mexico: CIMMyT and IRRI.

McLaren, C. G., R. Bruskiewich, A.M. Portugal, and A.B. Cosico. 2005. The International Rice Information System. 
A platform for meta-analysis of rice crop data. \emph{Plant Physiology} \bold{139}: 637-642.
}
\keyword{datasets}
