\name{wheat.sets}
\alias{wheat.sets}
\title{Sets for cross validation (CV)}
\description{
Is a vector (599 x 1) that assigns observations to 10
disjoint sets; the assignment was generated at random.
This is used later to conduct a 10-fold CV.
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\keyword{datasets}
