\name{wheat.Y}
\alias{wheat.Y}
\title{Grain yield}
\description{
A matrix (599 x 4) containing the 2-yr average grain
yield of each of these lines in each of the four environments
(phenotypes were standardized to a unit variance within each environment).
}

\source{
  International Maize and Wheat Improvement Center (CIMMYT), Mexico.
}

\keyword{datasets}
