\name{mice.pheno}
\alias{mice.pheno}
\title{Phenotypical data for the mice dataset}
\description{

A data frame with pheotypical information related to diabetes.
The data frame has several columns: SUBJECT.NAME, PROJECT.NAME, PHENOTYPE.NAME,
Obesity.BMI, Obesity.BodyLength, Date.Month,
Date.Year, Date.Season,cDate.StudyStartSeconds,
Date.Hour, Date.StudyDay, GENDER,
EndNormalBW, CoatColour, CageDensity,
Litter, cage.

The phenotypes are described in http://gscan.well.ox.ac.uk.
}

\source{
 \url{http://gscan.well.ox.ac.uk}
}

\keyword{datasets}
